整理路徑,分類為cb,reop
