+++ logs/myapp-2023-10-20.log
@@ -1,2 +1,1 @@
-2023-10-20 11:17:22,452 INFO Abc [main] aaa is aaa
-2023-10-20 11:17:47,474 INFO Abc [main] aaa is aaa
+2023-10-20 17:45:29,663 INFO Application [main] aaa is aaa
