+++ src/main/java/cb/PlatformAccount.java
@@ -1,4 +1,4 @@
-package controller;
+package cb;
 
 import java.util.Map;
