+++ src/main/java/cb/TransferProxyLine.java
@@ -1,4 +1,4 @@
-package controller;
+package cb;
 
 import java.util.Map;
