+++ src/main/java/controller/CBServlet.java
@@ -1,5 +1,9 @@
 package controller;
 
+import cb.MemberAccount;
+import cb.PlatformAccount;
+import cb.ProxyLine;
+import cb.TransferProxyLine;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import tool.GetIp;
@@ -16,42 +20,13 @@ import java.io.PrintWriter;
 import java.util.HashMap;
 import java.util.Map;
 
-@WebServlet("/proxy")//打在網頁根目錄後
-public class HelloWorldServlet extends HttpServlet {
-    private static final Logger logger = LogManager.getLogger(HelloWorldServlet.class);
+@WebServlet("/cb")//打在網頁根目錄後
+public class CBServlet extends HttpServlet {
+    private static final Logger logger = LogManager.getLogger(CBServlet.class);
     TransferProxyLine transferProxyLine = new TransferProxyLine();
     ProxyLine proxyLine = new ProxyLine();
-
     PlatformAccount platformAccount = new PlatformAccount();
     MemberAccount memberAccount = new MemberAccount();
-    GetIp getIp = new GetIp();
-    SomeCookie someCookie = new SomeCookie();
-
-    /**
-     * ip
-     * @param request
-     * @param response
-     * @throws IOException
-     */
-    private void isme(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        PrintWriter out = response.getWriter();
-        String password = request.getParameter("password").trim();
-        if (password.equals("qwe123")) {
-            // 比對本地
-//            if (!"10.252.1.149".equals(getIp.getLocalIP())) {
-//                out.write("ipno");
-//            } else {
-                logger.info("hello ip is {}",getIp.getLocalIP());
-                someCookie.setCookie("isme", getIp.getLocalIP(), 600, response);
-                response.setHeader("Refresh", "1;URL=home.jsp");
-                out.write("sus");
-//            }
-        } else {
-            out.write("run");
-        }
-        out.close();
-    }
-
     private void memberAccountFreeze(HttpServletRequest request, HttpServletResponse response) throws IOException {
         request.setCharacterEncoding("UTF-8");//对返回浏览器数据没啥用，不过建议添加
         response.setHeader("Content-type", "text/html;charset=UTF-8");//告知浏览器编码方式;
@@ -267,55 +242,6 @@ public class HelloWorldServlet extends HttpServlet {
         out.close();
     }
 
-
-    /**
-     * int page, int pageSize,
-     * String createDatTime, String createDatTimeEnd
-     * to
-     * map.put("page",page);
-     * map.put("createDatTime",createDatTime);
-     * map.put("createDatTimeEnd",createDatTimeEnd);
-     * 將參數統一放在map，在別支方法由map取出
-     */
-    private void param2Map2paramStr(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        String paramString = request.getParameter("adjustStr");
-        StringBuffer stringBuffer = new StringBuffer();
-        StringBuffer stringBuffer2 = new StringBuffer();
-        stringBuffer.append("Map<String, Object> map = new HashMap<>();<br>");
-        String type = "", parameter = "";
-//        System.out.println(paramString);
-        if (paramString.contains(",")) {
-//            System.out.println(",>"+paramString);
-            for (String s : paramString.split(",")) {
-//                System.out.println("s>"+s);
-                // 去頭尾空白
-                if(!s.trim().equals("")){
-                    type = s.trim().split(" ")[0];
-                    parameter = s.trim().split(" ")[1];
-                }
-                stringBuffer.append("map.put(\"" + parameter + "\" ," + parameter + ");<br>");
-                stringBuffer2.append(type + " " + parameter + " = (" + type + ") map.get(\"" + parameter + "\");<br>");
-            }
-        } else if (paramString.contains(";")) {
-//            System.out.println(";>"+paramString);
-            String str1 = "";
-            for (String s : paramString.split(";")) {
-                // 去頭尾空白
-                if(!s.trim().equals("")){
-                    str1 = s.trim().split("=")[0];
-                    type = str1.trim().split(" ")[0];
-                    parameter = str1.trim().split(" ")[1];
-                }
-                stringBuffer.append("map.put(\"" + parameter + "\" ," + parameter + ");<br>");
-                stringBuffer2.append(type + " " + parameter + " = (" + type + ") map.get(\"" + parameter + "\");<br>");
-            }
-        }
-
-        PrintWriter out = response.getWriter();
-        out.write(stringBuffer.toString() + "<br>" + stringBuffer2.toString());
-        out.close();
-    }
-
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         doPost(request, response);
@@ -324,13 +250,9 @@ public class HelloWorldServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String step = request.getParameter("step");
-        logger.info("doPost step is {}",step);
+        logger.info("CBServlet: {}",step);
         if (step != null) {
-            if (step.equals("isme")) {
-                isme(request, response);
-            } else if (step.equals("adjust")) {
-                param2Map2paramStr(request, response);
-            } else if (step.contains("platformAccountFreeze")) {
+            if (step.contains("platformAccountFreeze")) {
                 platformAccountFreeze(request, response);
             } else if (step.contains("memberAccountFreeze")) {
                 memberAccountFreeze(request, response);
@@ -339,9 +261,6 @@ public class HelloWorldServlet extends HttpServlet {
             } else if (step.contains("transferProxyLine")) {
                 transferProxyLine(request, response);
             }
-        } else {
-            RequestDispatcher dispatcher = request.getRequestDispatcher("/home.jsp");
-            dispatcher.forward(request, response);
         }
     }
 }
