+++ src/main/webapp/cb/js/memberAccount.js
@@ -49,7 +49,7 @@ $("#MAfreeze1").click(function(){
     if(dataCheck(data)){
         $.ajax({
             type: "POST",
-            url: "/proxy",
+            url: "/cb",
             data: data,
             dataType: "HTML",
             success: function(result){
