+++ src/main/webapp/css/home.css
@@ -1,0 +1,10 @@
+.border{
+  border: 2px solid black;
+  margin-left:20px;
+}
+.sql-code {
+    background-color: #f2f2f2;
+    padding: 10px;
+    font-family: monospace;
+    border: 1px solid #ccc;
+}
