--- src/main/webapp/css/transferProxyLine.css
@@ -1,10 +1,0 @@
-.border{
-      border: 2px solid black;
-      margin-left:20px;
-  }
-    .sql-code {
-        background-color: #f2f2f2;
-        padding: 10px;
-        font-family: monospace;
-        border: 1px solid #ccc;
-    }
