+++ src/main/webapp/home.jsp
@@ -13,6 +13,7 @@
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
     <script src="js/openTabs.js"></script>
     <script src="js/home.js"></script>
+    <link  rel="stylesheet" type="text/css" href="css/home.css">
     <link  rel="stylesheet" type="text/css" href="css/tabs.css">
     <title>stopitgetsomehelp</title>
 </head>
@@ -27,9 +28,9 @@ boolean isme = someCookie.checkCookie(request, response);
 <%}else{%>
     <jsp:include page="personal.jsp" />
 <%}%>
-<jsp:include page="transferProxyLine.jsp" />
-<jsp:include page="platformAccount.jsp" />
-<jsp:include page="memberAccount.jsp" />
+<jsp:include page="cb/transferProxyLine.jsp" />
+<jsp:include page="cb/platformAccount.jsp" />
+<jsp:include page="cb/memberAccount.jsp" />
 <jsp:include page="codeAdjust.jsp" />
 </body>
 </html>
